refactor config validation
