+++ packages/expo-cli/src/commands/utils/__tests__/validateApplicationId-test.ts
@@ -1,4 +1,4 @@
-import { validateBundleId } from '../ConfigValidation';
+import { validateBundleId } from '../validateApplicationId';
 
 describe(validateBundleId, () => {
   it(`validates`, () => {
